--- a/promql/engine.go
+++ b/promql/engine.go
@@ -24,7 +24,6 @@ import (
 	"runtime"
 	"sort"
 	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -1388,12 +1387,7 @@ func (ev *evaluator) eval(expr parser.Expr) (parser.Value, annotations.Annotatio
 		}, e.Param, e.Expr)
 
 	case *parser.Call:
-		call, ok := FunctionCalls[e.Func.Name]
-		if !ok && strings.HasPrefix(e.Func.Name, "udf_") {
-			call = genFunctionCall(e.Func.Name)
-			FunctionCalls[e.Func.Name] = call
-		}
-
+		call := FunctionCalls[e.Func.Name]
 		if e.Func.Name == "timestamp" {
 			// Matrix evaluation always returns the evaluation time,
 			// so this function needs special handling when given