--- a/promql/parser/functions.go
+++ b/promql/parser/functions.go
@@ -396,6 +396,17 @@ var Functions = map[string]*Function{
 	},
 }
 
+func AddFunction(name string) {
+	if _, ok := Functions[name]; ok {
+		return
+	}
+	Functions[name] = &Function{
+		Name:       name,
+		ArgTypes:   []ValueType{ValueTypeMatrix},
+		ReturnType: ValueTypeVector,
+	}
+}
+
 // getFunction returns a predefined Function object for the given name.
 func getFunction(name string, functions map[string]*Function) (*Function, bool) {
 	function, ok := functions[name]