This PR adds User Defined Functions (UDFs) to PromQL.

Why use UDFs?
- Users have asked for functions that we are reluctant to include, e.g. [mad_over_time](https://github.com/prometheus/prometheus/issues/5514) (Median absolute deviation, similar to standard deviation but ignores outliers)
- Similar products like VictoriaMetrics have implemented these functions or something similar
- It'd be good to give users the flexibility to define their own functions, without having to make the case for it being needed and waiting for it to be added to Prometheus, and so they can use their preferred variety (e.g. there are different kinds of MAD functions)
- We would not need to support a lot of additional functions, and we would not have a lot of rarely-used functions to confuse users

Limitations of the current implementation:
- Slower than native functions (see below benchmarks)
- Adds a dependency on a new library (which is not updated that often)
- Allowing scripting may add security risks
- Uses a lesser known scripting language, but one that is simple and based on go ([Tengo](https://github.com/d5/tengo))
- UDFs won't appear in query autocomplete
- UDFs are shared across the whole instance, not separated by tenant
- Cannot aggregate multiple series
- Input and output types are limited, but they cover most of the functions that we currently have
- Does not support native histograms yet, but this is theoretically possible to add though I haven't tested it

Example YML (add to main Prometheus config):
```
udfs:
  - name: mad_over_time
    input_types:
      - matrix
    modules:
      - math
    util: true
    src: |
      math := import("math")
      util := import("util")

      mad := func(seq) {
        median := util.median(seq)
        if is_error(median) {
          return error("cannot compute median")
        }
        deltas := []
        for x in seq {
          deltas = append(deltas, math.abs(x-median))
        }
        return util.median(deltas)
      }

      output := mad(input0)
```

Benchmark results (before is native function, after is UDF):
```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                  │ BenchmarkStdDevOverTime_before.txt │  BenchmarkStdDevOverTime_after.txt   │
                  │               sec/op               │   sec/op     vs base                 │
StdDevOverTime-16                          16.22µ ± 1%   52.71µ ± 3%  +224.97% (p=0.000 n=10)

                  │ BenchmarkStdDevOverTime_before.txt │   BenchmarkStdDevOverTime_after.txt    │
                  │                B/op                │     B/op       vs base                 │
StdDevOverTime-16                         10.32Ki ± 0%   109.33Ki ± 0%  +959.63% (p=0.000 n=10)

                  │ BenchmarkStdDevOverTime_before.txt │  BenchmarkStdDevOverTime_after.txt   │
                  │             allocs/op              │  allocs/op   vs base                 │
StdDevOverTime-16                           145.0 ± 0%   1056.0 ± 0%  +628.28% (p=0.000 n=10)

goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/promql
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
               │ BenchmarkMadOverTime_before.txt │     BenchmarkMadOverTime_after.txt     │
               │             sec/op              │    sec/op     vs base                  │
MadOverTime-16                       26.15µ ± 3%   288.25µ ± 4%  +1002.11% (p=0.000 n=10)

               │ BenchmarkMadOverTime_before.txt │     BenchmarkMadOverTime_after.txt      │
               │              B/op               │     B/op       vs base                  │
MadOverTime-16                      19.87Ki ± 0%   245.13Ki ± 0%  +1133.71% (p=0.000 n=10)

               │ BenchmarkMadOverTime_before.txt │    BenchmarkMadOverTime_after.txt     │
               │            allocs/op            │  allocs/op   vs base                  │
MadOverTime-16                        149.0 ± 0%   4052.0 ± 0%  +2619.46% (p=0.000 n=10)
```
BenchmarkStdDevOverTime is a better comparison as the native function and UDF are implemented in very similar ways. For BenchmarkMadOverTime, the implementations are very different, and there is room to optimise the UDF performance by using more efficient methods to calculate the median.