@bboreham 2023-09-20
I think you should set out your requirements and proposal in a GitHub issue, then link this PR to it.
That way we can debate requirements and solution separately, and it might be that there are multiple proposed solutions.

Also, would this help with #11609?