@beorn7 2023-09-20
I think #11609 is aiming for adding new concepts to PromQL itself, while this PR uses a different language (Tengo) to add new functions to PromQL (which you could already do easily in Go if you are fine with compiling your own Prometheus code – this PR just allows to "script" functions in a Go-like language and run them within an already compiled standard Prometheus.)