@zenador 2023-09-20
@beorn7 Sure, do you mean discuss in a meeting, on a Github issue/PR or at the dev summit?

I chose Tengo for this implementation as their benchmarks indicate that it is faster than the alternatives (including gpython), though I didn't check the benchmarks closely.

@bboreham Embedding a scripting language is something that has been discussed previously (it's also a dev summit item) and I think others have more context on this. I couldn't find an issue though, so please let me know if I should create one @beorn7.

I think it can help with https://github.com/prometheus/prometheus/issues/11609 but not directly, in that it doesn't do exactly as requested but it should be able to achieve the goals in another way. It doesn't allow defining variables in PromQL as requested there, but you can define variables in Tengo for a new UDF (so the UDF code should be more readable), and you can reuse the UDF in multiple places, so queries should be shorter and more understandable overall.

As mentioned in the dev summit item, it would help with issues like https://github.com/prometheus/prometheus/issues/5514 https://github.com/prometheus/prometheus/pull/12497/ and https://github.com/prometheus/prometheus/pull/9016 where users want to use new functions, but Prometheus wants to be selective about the officially supported functions that need to be maintained. This way, users can add any functions they like without needing it to be accepted by Prometheus first (and it seems very difficult to push for new functions to be accepted), and functions can have many varieties (see MAD and the [discussion on holt winters](https://github.com/prometheus/prometheus/issues/2458)), so users can choose which variety they want to use for themselves.