@beorn7 2023-09-20
Since the dev summit is literally next week, I would say we can just wait to see the topic discussed and take it from there.