@beorn7 2023-10-04
Unfortunately, the dev summit didn't manage to discuss the topic relevant to this PR. It will still happen eventually on the monthly online summits, but maybe we can at least get an idea how to proceed by discussing it here.

@juliusv @roidelapluie you two might be the most relevant custodians of PromQL, so flagging you (but I don't want to exclude anyone else, please speak up). What do you generally think about this approach? Personally, I think this is probably worth a design doc, but your input would good to find out about the direction to take in the doc. Or worst case, if you are fundamentally against this, we could discuss the reasons before investing a lot of effort into writing a design doc.