@beorn7 2023-10-31
We discussed this PR and the topics around it at the recent dev summit. I would like to summarize the relevant outcome here.

Generally, we didn't come to a consensus that we want this. However, there was a fairly easy consensus that more functions should be accepted into PromQL behind a feature flag. The bar was set to "at least somewhat generally useful" and "cannot be easily modeled by existing PromQL primitives". Since the new functions will be behind a feature flag, there is no danger of a "second holt_winters" (where a function was added under that name that isn't the "real" Holt Winters and was also of limited use, but now we have to maintain it forever). New functions can then be graduated to stable (or removed again) depending on the experience made in the wild.

Adding native functions more easily would reduce the weight of one of the basic motivations behind this PR. Which in turn would increase the relative weight of the concerns.

While performance concerns were mentioned, the main concerns were more about operational aspects, like the following:
- It is harder to share PromQL solutions if you also have to hand around your UDFs. Instead of a few "canonical" solutions, we'll breed a whole zoo of solutions involving different UDFs.
- While many will probably use UDFs for more readable PromQL expression, UDFs also hide what's actually going on, so in practice (and in the heat of an outage) it might actually be much harder to understand what a PromQL expression is doing if it involves UDFs.
- Many different approaches for the same problem were discussed, and there is little clarity which one is best to pursue (plugins on the binary level, implement UDFs through a remote-read backend, allow some kind of UDFs within PromQL, …).

In favor of this approach, people mentioned the ease of adding new functions for problems at hand and enabling quick experimentation. This could also be used to see what kind of functions users implement most often, so that we can take that as an inspiration for future built-in PromQL functions.

To avoid name collisions, UDFs should come with some unique prefix (so that the later introduction of built-in functions of the same name doesn't cause an error) or maybe UDFs can be called through a special built-in PromQL function (e.g. `call(my_own_mad_function, my_metric{foo="bar"})`.

Overall, I would recommend to wait how more openness towards adding native functions plays out. Right now, it will probably be a lot of effort to get sufficient buy-in for UDFs via an embedded scripting engine.