@zenador 2023-10-31
Noted, thanks for the summary!

Just an FYI, this PR already avoids name collisions by enforcing a `udf_` prefix.