@beorn7 2024-08-27
mad_over_time is implemented, which kind of proves the point that being more liberal with adding new functions helps. https://github.com/prometheus/prometheus/pull/12503 is another new function that got merged. 

A counter example (i.e. where we are still reluctant to "just merge") is https://github.com/prometheus/prometheus/pull/14440 , but I believe this is for good reasons, i.e. implementing what the user wants is actually hard. So having an option to add UDFs in a simple way would just add to the confusion (e.g. having a lot of half-bred solutions floating around).

In the spirit of not keeping issues open forever on which we do not plan to work, let's close this now, but keep it in mind as a POC in case we circle back and want UDFs after all.

@zenador thanks again for this work. It was definitely useful (and might prove useful in the future).